backend/requirements.txt
django
djangorestframework
pytesseract
pillow
transformers
urllib3
tzdata
typing-extenstions
sqlparse
safetensors
regex
pyyaml
packaging
numpy
idna
fsspec
filelock
colorama
charset-normalizer
certifi
asgiref
tqdm
requests
tokenizers
huggingface-hub
rake-nltk
pdfplumber
tf-keras